+++ cpu/pfsolve.h
@@ -512,16 +512,16 @@ namespace pFROST {
 		}
 		inline void		countAll			() {
 			std::atomic<uint32> cls = 0, lits = 0;
-			std::atomic<int> threadIdx = 0;
-			uint32 batchSize = scnf.size() / workerPool.count();
-			uint32 remainder = scnf.size() % workerPool.count();
+			std::atomic<uint32> threadIdx = 0;
+			uint32 batchSize = (uint32)scnf.size() / workerPool.count();
+			uint32 remainder = (uint32)scnf.size() % workerPool.count();
 
 			workerPool.doWork([&] {
-				int tid = threadIdx++;
+				uint32 tid = threadIdx++;
 				uint32 c = 0;
 				uint32 l = 0;
 
-				uint32 begin = batchSize * tid + std::min((uint32)tid, remainder);
+				uint32 begin = batchSize * tid + std::min(tid, remainder);
 				uint32 end = begin + batchSize + (tid < remainder ? 1 : 0);
 				
 				for (uint32 i = begin; i < end; i++) {
